Messages
- Error Messages added
- Option Messages added
- New Category in Changelog called "replaced"
- Dialog will pop up if a file already exist in the destination folder
- Some Code changes

+++ CleanDesktop/src/Functions/ReadDesktop.java
@@ -10,7 +10,9 @@ public class ReadDesktop {
 	String path;
 	int allowedDeleteSize = 22;
 	ReadSettingsAndGetCategory gcap = new ReadSettingsAndGetCategory();
-	
+	Messages messages = new Messages();
+	String [] savedFilesForChangelog;
+ 	
 	public ReadDesktop(String prepath) {
 		path = prepath;
 		reading();
@@ -19,36 +21,87 @@ public class ReadDesktop {
 	public void reading() {
 		File desktopfile = new File(desktoppath);
 		String filelist[] = desktopfile.list();
-		String transferedFiles [] = new String [filelist.length];
+		savedFilesForChangelog = new String [filelist.length];
 		//String transferedFilesDestinations [] = new String [filelist.length];
 		int counter = filelist.length;
 		for(int i = 0; i <= counter - 1; i++) {			
 			File file = new File (desktoppath + "\\" + filelist[i]);
-			long size = file.length(); //Funktioniert nicht überall
-			//int size = (int) file.length();
+			long size = file.length();
 			getDeleteSize();
 			
-			if (!Whitelist.isFileInWhitelist(file.getName()) && file.getName().contains(".")) {
+			if (!Whitelist.isFileInWhitelist(file.getName()) && file.getName().contains(".") && file.isFile()) {
 				if (Boolean.parseBoolean(gcap.getSetup("del_customFiles")) == true) {
 					if (size > allowedDeleteSize) {
-						transferedFiles[i] = file.getName();
-						new TransferFiles(filelist[i]); 
+						transferFiles(file.getName(), i);
 					}else {
-					transferedFiles[i] = "del_"+file.getName();
+						savedFilesForChangelog[i] = "del_"+file.getName();
 					file.delete();	
 					}
 				}else if (size > 1) {
-					transferedFiles[i] = file.getName();
-					new TransferFiles(filelist[i]); 
-					
+					transferFiles(file.getName(), i);
 				}else {
-					transferedFiles[i] = "del_"+file.getName();
+					savedFilesForChangelog[i] = "del_"+file.getName();
 					file.delete();
 				}
 			}
 		
 		}
-		new GenerateChangelog(transferedFiles);
+		new GenerateChangelog(savedFilesForChangelog);
+	}
+	
+	public void transferFiles(String filename, int counter) {
+		if(gcap.selectCategory(filename) == 1) {
+			if (!checkIfFileExists(gcap.getSetup("img_destination"), filename)) {
+				new TransferFiles(filename, gcap.getSetup("img_destination"));
+				savedFilesForChangelog[counter] = filename;
+			}else if (checkIfUserWantToReplace(gcap.getSetup("img_destination"), filename)) {
+				new TransferFiles(filename, gcap.getSetup("img_destination"));
+				savedFilesForChangelog[counter] = "rep_"+filename;
+			}
+		}else if (gcap.selectCategory(filename) == 2) {
+			if (!checkIfFileExists(gcap.getSetup("msc_destination"), filename)) {
+				new TransferFiles(filename, gcap.getSetup("msc_destination"));
+				savedFilesForChangelog[counter] = filename;
+			}else if (checkIfUserWantToReplace(gcap.getSetup("img_destination"), filename)) {
+				new TransferFiles(filename, gcap.getSetup("img_destination"));
+				savedFilesForChangelog[counter] = "rep_"+filename;
+			}
+			
+		}else if (gcap.selectCategory(filename) == 3) {
+			if (!checkIfFileExists(gcap.getSetup("vid_destination"), filename)) {
+				new TransferFiles(filename, gcap.getSetup("vid_destination"));
+				savedFilesForChangelog[counter] = filename;
+			}else if (checkIfUserWantToReplace(gcap.getSetup("img_destination"), filename)) {
+				new TransferFiles(filename, gcap.getSetup("img_destination"));
+				savedFilesForChangelog[counter] = "rep_"+filename;
+			}
+		}else {
+			if (!checkIfFileExists(gcap.getSetup("doc_destination"), filename)) {
+				new TransferFiles(filename, gcap.getSetup("doc_destination"));
+				savedFilesForChangelog[counter] = filename;
+			}else if (checkIfUserWantToReplace(gcap.getSetup("img_destination"), filename)) {
+				new TransferFiles(filename, gcap.getSetup("img_destination"));
+				savedFilesForChangelog[counter] = "rep_"+filename;
+			}
+		}
+	}
+	
+	private boolean checkIfFileExists(String destination, String filename) {
+		File file = new File(destination+"/"+filename);
+ 		if (file.exists()) {
+				return true;
+		}
+ 		return false;
+	}
+	
+	private boolean checkIfUserWantToReplace(String destination, String filename) {
+		Object [] options = {"Replace","Skip"};
+		File file = new File(destination+"/"+filename);
+		if (messages.optionMessages((byte) 1, filename, options) == 0) {
+			file.delete();
+			return true;
+		}
+		return false;
 	}
 	
 	public void getDeleteSize() {
@@ -60,5 +113,4 @@ public class ReadDesktop {
 			allowedDeleteSize = Integer.parseInt(gcap.getSetup("del_customFilesRange"))*1000000000;
 		}
 	}
-
 }
